+++ rend3/src/lib.rs
@@ -55,6 +55,7 @@ pub mod util {
     pub mod bind_merge;
     pub mod buffer;
     pub mod frustum;
+    pub mod mipmap;
     pub mod output;
     pub mod registry;
     pub mod typedefs;