+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to cargo's version of [Semantic Versioning](https://sem
 ### Added
 - rend3: Materials now have a `Transparency` field that supports Opaque, Cutout, and Blend transparency modes.
 - rend3: `AlbedoComponent::TextureVertexValue` to represent `texture * vertex * constant`
+- rend3: Mipmaps can be generated automatically on the gpu without the user needing to upload them.
 - rend3 & rend3-pbr: Use `wgpu-profiler` to generate GPU timings that show up as `RendererStatistics`.
 - rend3 & rend3-pbr: Annotate most code with `profiling` spans.
 - rend3-pbr: All major rendering spans are labeled and show up in renderdoc
@@ -34,6 +35,7 @@ and this project adheres to cargo's version of [Semantic Versioning](https://sem
   - All resources are removed the `render()` after they are deleted.
 - rend3: `Texture::width` and `Texture::height` replaced with `Texture::size`
 - rend3: `RendererStatistics` is now an alias for `Vec<wgpu_profiler::GpuTimerScopeResult>`
+- rend3: `Texture::mip_levels` was split into `mip_count` and `mip_source` allowing you to easily auto-generate mipmaps.
 - rend3-pbr: creation and resizing's `resolution` argument replaced with options containing resolution and sample count.
   
 ### Updated